trying to finalize the cleanup of thumbnail optimizations. fixes the optimized lookup of tabular data files; proper thumbnail generation for worldmap files; isolates the optimized thumbnail handling in its own ViewScope wrapper; makes the dataset page call the wrapper to obtain the dataset-level logo for the top of the page, when in optimized mode; properly implements the persistence of auto-selected thumbnails.
