+++ src/main/java/edu/harvard/iq/dataverse/search/SearchIncludeFragment.java
@@ -15,6 +15,7 @@ import edu.harvard.iq.dataverse.DvObjectServiceBean;
 import edu.harvard.iq.dataverse.PermissionServiceBean;
 import edu.harvard.iq.dataverse.PermissionsWrapper;
 import edu.harvard.iq.dataverse.SettingsWrapper;
+import edu.harvard.iq.dataverse.ThumbnailServiceWrapper;
 import edu.harvard.iq.dataverse.WidgetWrapper;
 import edu.harvard.iq.dataverse.dataaccess.ImageThumbConverter;
 import edu.harvard.iq.dataverse.dataset.DatasetThumbnail;
@@ -72,6 +73,8 @@ public class SearchIncludeFragment implements java.io.Serializable {
     SettingsWrapper settingsWrapper;
     @Inject
     PermissionsWrapper permissionsWrapper;
+    @Inject
+    ThumbnailServiceWrapper thumbnailServiceWrapper;
     @Inject
     WidgetWrapper widgetWrapper;  
     @EJB
@@ -1159,9 +1162,7 @@ public class SearchIncludeFragment implements java.io.Serializable {
     }
 
     public void setDisplayCardValues() {
-        int i = 0;
-        dvobjectThumbnailsMap = new HashMap<>();
-        dvobjectViewMap = new HashMap<>();
+
         Set<Long> harvestedDatasetIds = null;
         for (SolrSearchResult result : searchResultsList) {
             //logger.info("checking DisplayImage for the search result " + i++);
@@ -1173,10 +1174,10 @@ public class SearchIncludeFragment implements java.io.Serializable {
                  * downloadable image -
                  * https://github.com/IQSS/dataverse/issues/3616
                  */
-                result.setImageUrl(getDataverseCardImageUrl(result));
+                result.setImageUrl(thumbnailServiceWrapper.getDataverseCardImageAsBase64Url(result));
             } else if (result.getType().equals("datasets")) {
                 
-                result.setImageUrl(getDatasetCardImageUrl(result));
+                result.setImageUrl(thumbnailServiceWrapper.getDatasetCardImageAsBase64Url(result));
                 
                 if (result.isHarvested()) {
                     if (harvestedDatasetIds == null) {
@@ -1185,7 +1186,7 @@ public class SearchIncludeFragment implements java.io.Serializable {
                     harvestedDatasetIds.add(result.getEntityId());
                 }
             } else if (result.getType().equals("files")) {
-                result.setImageUrl(getFileCardImageUrl(result));
+                result.setImageUrl(thumbnailServiceWrapper.getFileCardImageAsBase64Url(result));
                 if (result.isHarvested()) {
                     if (harvestedDatasetIds == null) {
                         harvestedDatasetIds = new HashSet<>();
@@ -1194,8 +1195,8 @@ public class SearchIncludeFragment implements java.io.Serializable {
                 }
             }
         }
-        dvobjectThumbnailsMap = null;
-        dvobjectViewMap = null;
+        
+        thumbnailServiceWrapper.resetObjectMaps();
         
         // Now, make another pass, and add the remote archive descriptions to the 
         // harvested dataset and datafile cards (at the expense of one extra 
@@ -1260,259 +1261,7 @@ public class SearchIncludeFragment implements java.io.Serializable {
         }
         
     }
-
-    private Map<Long, String> dvobjectThumbnailsMap = null;
-    private Map<Long, DvObject> dvobjectViewMap = null;
-
-    private String getAssignedDatasetImage(Dataset dataset) {
-        if (dataset == null) {
-            return null;
-        }
-
-        DataFile assignedThumbnailFile = dataset.getThumbnailFile();
-
-        if (assignedThumbnailFile != null) {
-            Long assignedThumbnailFileId = assignedThumbnailFile.getId();
-
-            if (this.dvobjectThumbnailsMap.containsKey(assignedThumbnailFileId)) {
-                // Yes, return previous answer
-                //logger.info("using cached result for ... "+assignedThumbnailFileId);
-                if (!"".equals(this.dvobjectThumbnailsMap.get(assignedThumbnailFileId))) {
-                    return this.dvobjectThumbnailsMap.get(assignedThumbnailFileId);
-                }
-                return null;
-            }
-
-            String imageSourceBase64 = ImageThumbConverter.getImageThumbnailAsBase64(
-                    assignedThumbnailFile,
-                    ImageThumbConverter.DEFAULT_CARDIMAGE_SIZE);
-
-            if (imageSourceBase64 != null) {
-                this.dvobjectThumbnailsMap.put(assignedThumbnailFileId, imageSourceBase64);
-                return imageSourceBase64;
-            }
-
-            // OK - we can't use this "assigned" image, because of permissions, or because 
-            // the thumbnail failed to generate, etc... in this case we'll 
-            // mark this dataset in the lookup map - so that we don't have to
-            // do all these lookups again...
-            this.dvobjectThumbnailsMap.put(assignedThumbnailFileId, "");
-            
-            // TODO: (?)
-            // do we need to cache this datafile object in the view map?
-            // -- L.A., 4.2.2
-        }
-
-        return null;
-
-    }
-
-    // it's the responsibility of the user - to make sure the search result
-    // passed to this method is of the Datafile type!
-    private String getFileCardImageUrl(SolrSearchResult result) {
-        // Before we do anything else, check if it's a harvested dataset; 
-        // no need to check anything else if so (harvested objects never have 
-        // thumbnails)
-        
-        if (result.isHarvested()) {
-            return null; 
-        }
-        
-        Long imageFileId = result.getEntity().getId();
-
-        if (imageFileId != null) {
-            if (this.dvobjectThumbnailsMap.containsKey(imageFileId)) {
-                // Yes, return previous answer
-                //logger.info("using cached result for ... "+datasetId);
-                if (!"".equals(this.dvobjectThumbnailsMap.get(imageFileId))) {
-                    return this.dvobjectThumbnailsMap.get(imageFileId);
-                }
-                return null;
-            }
-
-            String cardImageUrl = null;
-            
-            if (result.getTabularDataTags() != null) {
-                for (String tabularTagLabel : result.getTabularDataTags()) {
-                    DataFileTag tag = new DataFileTag();
-                    try {
-                        tag.setTypeByLabel(tabularTagLabel);
-                        tag.setDataFile((DataFile) result.getEntity());
-                        ((DataFile) result.getEntity()).addTag(tag);
-                    } catch (IllegalArgumentException iax) {
-                        // ignore 
-                    }
-                }
-            }
-
-            if ((!((DataFile)result.getEntity()).isRestricted()
-                        || permissionsWrapper.hasDownloadFilePermission(result.getEntity()))
-                    && dataFileService.isThumbnailAvailable((DataFile) result.getEntity())) {
-                
-                cardImageUrl = ImageThumbConverter.getImageThumbnailAsBase64(
-                        (DataFile) result.getEntity(),
-                        ImageThumbConverter.DEFAULT_CARDIMAGE_SIZE);
-            }
-
-            if (cardImageUrl != null) {
-                this.dvobjectThumbnailsMap.put(imageFileId, cardImageUrl);
-                //logger.info("datafile id " + imageFileId + ", returning " + cardImageUrl);
-
-                if (!(dvobjectViewMap.containsKey(imageFileId)
-                        && dvobjectViewMap.get(imageFileId).isInstanceofDataFile())) {
-
-                    dvobjectViewMap.put(imageFileId, result.getEntity());
-
-                }
-
-                return cardImageUrl;
-            } else {
-                this.dvobjectThumbnailsMap.put(imageFileId, "");
-            }
-        }
-        return null;
-    }
-
-    // it's the responsibility of the user - to make sure the search result
-    // passed to this method is of the Dataset type!
-    private String getDatasetCardImageUrl(SolrSearchResult result) {
-        // Before we do anything else, check if it's a harvested dataset; 
-        // no need to check anything else if so (harvested datasets never have 
-        // thumbnails)
-        
-        if (result.isHarvested()) {
-            return null; 
-        }
-        
-        Long datasetId = result.getEntity().getId();
-
-        if (datasetId != null) {
-            if (this.dvobjectThumbnailsMap.containsKey(datasetId)) {
-                // Yes, return previous answer
-                // (at max, there could only be 2 cards for the same dataset
-                // on the page - the draft, and the published version; but it's 
-                // still nice to try and cache the result - especially if it's an
-                // uploaded logo - we don't want to read it off disk twice). 
-                
-                if (!"".equals(this.dvobjectThumbnailsMap.get(datasetId))) {
-                    return this.dvobjectThumbnailsMap.get(datasetId);
-                }
-                return null;
-            }
-        }
-
-        if (((Dataset)result.getEntity()).isUseGenericThumbnail()) {
-            this.dvobjectThumbnailsMap.put(result.getEntityId(), "");
-            return null; 
-        }
-        
-        String cardImageUrl = null;
-        
-        Path path = Paths.get(((Dataset)result.getEntity()).getFileSystemDirectory() + File.separator + DatasetUtil.datasetLogoThumbnail + DatasetUtil.thumb48addedByImageThumbConverter);
-        if (Files.exists(path)) {
-            try {
-                byte[] bytes = Files.readAllBytes(path);
-                String base64image = Base64.getEncoder().encodeToString(bytes);
-                cardImageUrl = FileUtil.DATA_URI_SCHEME + base64image;
-                this.dvobjectThumbnailsMap.put(result.getEntityId(), cardImageUrl);
-                return cardImageUrl;
-            } catch (IOException ex) {
-                this.dvobjectThumbnailsMap.put(result.getEntityId(), "");
-                return null; 
-                // (alternatively, we could ignore the exception, and proceed with the 
-                // regular process of selecting the thumbnail from the available 
-                // image files - ?)
-            }
-        } 
-
-        if (result.getEntity() != null) {
-            cardImageUrl = this.getAssignedDatasetImage((Dataset) result.getEntity());
-
-            if (cardImageUrl != null) {
-                //logger.info("dataset id " + result.getEntity().getId() + " has a dedicated image assigned; returning " + cardImageUrl);
-                return cardImageUrl;
-            }
-        }
-
-        Long thumbnailImageFileId = datasetVersionService.getThumbnailByVersionId(result.getDatasetVersionId());
-
-        if (thumbnailImageFileId != null) {
-            //cardImageUrl = FILE_CARD_IMAGE_URL + thumbnailImageFileId;
-            if (this.dvobjectThumbnailsMap.containsKey(thumbnailImageFileId)) {
-                // Yes, return previous answer
-                //logger.info("using cached result for ... "+datasetId);
-                if (!"".equals(this.dvobjectThumbnailsMap.get(thumbnailImageFileId))) {
-                    return this.dvobjectThumbnailsMap.get(thumbnailImageFileId);
-                }
-                return null;
-            }
-
-            DataFile thumbnailImageFile = null;
-
-            if (dvobjectViewMap.containsKey(thumbnailImageFileId)
-                    && dvobjectViewMap.get(thumbnailImageFileId).isInstanceofDataFile()) {
-                thumbnailImageFile = (DataFile) dvobjectViewMap.get(thumbnailImageFileId);
-            } else {
-                thumbnailImageFile = dataFileService.findCheapAndEasy(thumbnailImageFileId);
-                if (thumbnailImageFile != null) {
-                    // TODO:
-                    // do we need this file on the map? - it may not even produce
-                    // a thumbnail!
-                    dvobjectViewMap.put(thumbnailImageFileId, thumbnailImageFile);
-                } else {
-                    this.dvobjectThumbnailsMap.put(thumbnailImageFileId, "");
-                    return null;
-                }
-            }
-
-            if (dataFileService.isThumbnailAvailable(thumbnailImageFile)) {
-                cardImageUrl = ImageThumbConverter.getImageThumbnailAsBase64(
-                        thumbnailImageFile,
-                        ImageThumbConverter.DEFAULT_CARDIMAGE_SIZE);
-            }
-
-            if (cardImageUrl != null) {
-                this.dvobjectThumbnailsMap.put(thumbnailImageFileId, cardImageUrl);
-            } else {
-                this.dvobjectThumbnailsMap.put(thumbnailImageFileId, "");
-            }
-        }
-
-        //logger.info("dataset id " + result.getEntityId() + ", returning " + cardImageUrl);
-
-        return cardImageUrl;
-    }
     
-    // it's the responsibility of the user - to make sure the search result
-    // passed to this method is of the Dataverse type!
-    private String getDataverseCardImageUrl(SolrSearchResult result) {
-        return dataverseService.getDataverseLogoThumbnailAsBase64ById(result.getEntityId());
-    }
-    
-    /* 
-        These commented out methods below are old optimizations that are no longer 
-        necessary, since there is now a more direct connection between a harvested
-        dataset and its HarvestingClient configuration. -- L.A. 4.5 
-    */
-    /*
-    private Map<Long, String> getHarvestedDatasetDescriptions() {
-        if (harvestedDatasetDescriptions != null) {
-            return harvestedDatasetDescriptions;
-        }
-        harvestedDatasetDescriptions = dataverseService.getAllHarvestedDataverseDescriptions();
-        return harvestedDataverseDescriptions;
-    }*/
-    
-    /*private boolean isHarvestedDataverse(Long id) {
-        return this.getHarvestedDataverseDescriptions().containsKey(id);
-    }
-
-    private String getHarvestingDataverseDescription(Long id) {
-        if (this.isHarvestedDataverse(id)) {
-            return this.getHarvestedDataverseDescriptions().get(id);
-        }
-        return null;
-    }*/
     public enum SortOrder {
 
         asc, desc
